pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "data-science-ai-team"
version = "0.1.0"
description = "AI-agent-based machine learning ecosystem using IBM watsonx + open-source frameworks"
requires-python = ">=3.11"
dependencies = [
    # Orchestration frameworks
    "crewai>=0.80.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-ibm>=0.3.0",
    "langchain-community>=0.3.0",
    # IBM Cloud / watsonx
    "ibm-watsonx-ai>=1.1.0",
    # API layer
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    # ML / Data Science
    "scikit-learn>=1.5.0",
    "imbalanced-learn>=0.12.0",
    "xgboost>=2.1.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "mlflow>=2.14.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
    "ipykernel>=6.29.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipywidgets>=8.1.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
]

[project.scripts]
ai-team = "data_science_ai_team.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/data_science_ai_team"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
